+++ library/src/jvmTest/kotlin/io/github/littlesurvival/fetch/FetchTest.kt
@@ -73,8 +73,10 @@ class FetchTest {
 
     @Test
     fun testThread(): Unit = runBlocking {
-        // 568055, 535612, 564532, 565033, 557223, 535057, 568285, 567394, 568356, 568493, 566241, 551879, 568644, 568760, 547803, 555740, 535802, 519989
-        val threadResult = client.fetchThreadById(ThreadId(519989), reverse = true, page = 1)
+        // 568055, 535612, 564532, 565033, 557223, 535057, 568285,
+        // 567394, 568356, 568493, 566241, 551879, 568644, 568760,
+        // 547803, 555740, 535802, 519989, 562708, 533586, 557845
+        val threadResult = client.fetchThreadById(ThreadId(557845), reverse = true, page = 1)
         if (threadResult is YamiboResult.Success) {
             println("post size : ${threadResult.value.posts.size}")
         }